+++ test/libsolidity/syntaxTests/types/mapping/delete_storage_struct_with_mapping.sol
@@ -1,0 +1,14 @@
+contract D {
+	struct Test {
+		mapping(address => uint) balances;
+	}
+
+	Test test;
+
+	constructor()
+	{
+		delete test;
+	}
+}
+// ----
+// TypeError 2811: (102-113): Type containing a (nested) mapping cannot be deleted.